requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1

# Database (PostgreSQL)
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.10

# Celery & Redis
celery==5.3.4
redis==5.2.0
django-celery-beat==2.5.0
django-celery-results==2.5.1

# WebSocket & Channels
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# API & HTTP
requests==2.31.0
httpx==0.25.2
django-filter==23.5

# Maps & Geolocation
geopy==2.4.1
folium==0.15.1

# Utilities
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3

# Development
django-debug-toolbar==4.2.0
ipython==8.18.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
Faker==22.0.0
freezegun==1.4.0
responses==0.24.1
